Cleaned up server files and split out server.cpp
